refactor(local-ci): extract state-path helpers to state_paths.py
R2-1 first cut. Move 18 path/lock/log helpers (state_dir, queue_path,
results_dir, logs_dir, ensure_state_dirs, target_log_path, etc.) from
local_ci.py into a sibling state_paths.py module so that future
extracted modules (queue, recovery, transport) can depend on a thin,
stable seam instead of importing from the 11k-line god module.

- local_ci.py: 11,285 → 11,202 lines (-83). The helpers re-export at
  module scope so test_local_ci.py's importlib loader and the live CLI
  both see the same symbols.
- state_paths.py: 137 lines, no behavioral change.
- SCRIPT_DIR is now defined in both modules (matching definition — same
  filesystem location). Resolved identically because both files live in
  tools/local-ci/.
- sys.path.insert at local_ci.py top handles the
  spec_from_file_location code path used by test_local_ci.py.

Tests: 213/213 local_ci tests pass.
